Refactor helper module
