+++ examples/mnist-pytorch/client/entrypoint
@@ -7,7 +7,7 @@ import docker
 import fire
 import torch
 
-from fedn.utils.helpers import get_helper, save_metadata, save_metrics
+from fedn.utils.helpers.helpers import get_helper, save_metadata, save_metrics
 
 HELPER_MODULE = 'numpyhelper'
 helper = get_helper(HELPER_MODULE)
