+++ fedn/fedn/utils/helpers/__init__.py
@@ -1,3 +1,3 @@
-""" The plugins package is responsible for loading model helper functions supporting different ML frameworks. The :class:`fedn.utils.plugins.helperbase.HelperBase` is
-an abstract class which user can implement their own helper functions to support different ML frameworks. """
+""" The helpers package is responsible for serialization/deserialization and numerics operations 
+for various machine learning frameworks. """
 # flake8: noqa
