+++ fedn/fedn/utils/helpers/helpers.py
@@ -10,7 +10,7 @@ def get_helper(helper_module_name):
     :param helper_module_name: The name of the helper plugin module.
     :type helper_module_name: str
     :return: A helper instance.
-    :rtype: class: `fedn.utils.helpers.HelperBase`
+    :rtype: class: `fedn.utils.helpers.helpers.HelperBase`
     """
     helper_plugin = HELPER_PLUGIN_PATH.format(helper_module_name)
     helper = importlib.import_module(helper_plugin)
